[2024-11-04] Donwload / 채팅 API 적용 및 모듈설치 / 결제관리 코드 수정 Chat.jsx : 채널톡 API 적용 / 모듈설치 / 연동확인 PayMentMethod.jsx : const payhistory -> const PayMentMethod 로 변경 Chak(backup).jsx : 이전 채팅 파일 백업 ( API 오류 발생시 롤백용 )
